+++ scripts/path.rb
